nlp-engine/requirements.txt
flask==3.0.3
flask-cors==4.0.1
sentence-transformers==3.0.1
faiss-cpu==1.9.0.post1
numpy==1.26.4
pandas==2.2.2
